% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decorate_shiny.R
\name{decorate_shiny}
\alias{decorate_shiny}
\title{Builds a \code{\link{decorated}} object from a code chunk}
\usage{
decorate_shiny(x, ...)
}
\arguments{
\item{x}{A string, containing a chunk label.}

\item{...}{Chunk options to pass along}
}
\value{
A \code{decorated} object.
}
\description{
(Aliased from \code{flair}; see that package for further use.)
}
