% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gild.R
\name{gild}
\alias{gild}
\title{Replaces parts of code with interactive Shiny inputs}
\usage{
gild(x, ...)
}
\arguments{
\item{x}{An object of class \code{something}}

\item{...}{Named strings telling which parts of the source code to replace
with shiny variables.}
}
\value{
An object of class \code{something}
}
\description{
Replaces parts of code with interactive Shiny inputs
}
