% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shindig.R
\name{shindig}
\alias{shindig}
\title{Writes server-side code and runs it in server context}
\usage{
shindig(x, render_type)
}
\arguments{
\item{x}{An object of type \code{decorated}}

\item{render_type}{A string containing the name of a Shiny render function.}
}
\value{
An object of class \code{decorated}
}
\description{
Writes server-side code and runs it in server context
}
